+++ tests/testthat/test_nearest_neighbor_kknn.R
@@ -20,6 +20,7 @@ quiet_ctrl <- fit_control(verbosity = 0, catch = TRUE)
 test_that('kknn execution', {
 
   skip_if_not_installed("kknn")
+  library(kknn)
 
   # continuous
   # expect no error
@@ -78,7 +79,7 @@ test_that('kknn prediction', {
 
   # nominal
   res_xy_nom <- fit_xy(
-    iris_basic,
+    iris_basic %>% set_mode("classification"),
     control = ctrl,
     x = iris[, c("Sepal.Length", "Petal.Width")],
     y = iris$Species
@@ -89,11 +90,11 @@ test_that('kknn prediction', {
     newdata = iris[1:5, c("Sepal.Length", "Petal.Width")]
   )
 
-  expect_equal(uni_pred_nom, predict_class(res_xy_nom, iris[1:5, c("Sepal.Length", "Petal.Width")]))
+  expect_equal(uni_pred_nom, parsnip:::predict_class(res_xy_nom, iris[1:5, c("Sepal.Length", "Petal.Width")]))
 
   # continuous - formula interface
   res_form <- fit(
-    iris_basic,
+    iris_basic %>% set_mode("regression"),
     Sepal.Length ~ log(Sepal.Width) + Species,
     data = iris,
     control = ctrl